requirements.txt
streamlit==1.23.1
azure-cognitiveservices-vision-computervision=0.9.0
pillow=9.4.0